Move RHVoice_synth_params and RHVoice_voice_info to RHVoice.py
